Trimester-4/MICROSERVICES/CIA3/services/admin_service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.22.0
SQLAlchemy>=1.4,<2.1
psycopg2-binary>=2.9
python-jose[cryptography]>=3.3
pydantic-settings>=2.2
python-dotenv>=1.0
email-validator>=2.0
grpcio>=1.63.0
